BUG/MEDIUM: acl: always accept match "found"
The pattern match "found" fails to parse on binary type samples. The
reason is that it presents itself as an integer type which bin cannot
be cast into. We must always accept this match since it validates
anything on input regardless of the type. Let's just relax the parser
to accept it.

This problem might also exist in 1.5.
[ note: it affects outputs of types ADDR, IPv6, BIN and METH here ]
(cherry picked from commit 9bb49f6906eafed0eb0fde01cd403089a502a101)