BUG/MINOR: stream: don't force retries if the server is DOWN
Arkadiy Kulev noticed that if a server is marked down while a connection
is being trying to establish, we still insist on performing retries on
the same server, which is absurd. Better perform the redispatch if we
already know the server is down. Because of this, it's likely that the
observe-l4 and sudden-death mechanisms are not optimal an cannot help
much the connection which was used to detect the problem.

The fix should be backported to 1.6 and 1.5 at least.
(cherry picked from commit 49008c157e762c7ed31f9c88ad54b6d6c42fbd77)
(cherry picked from commit d005682283c369c8170edb8ecb7e67450d9dcbd6)

[ the logic in 1.5 is slightly different since we don't have the
  redispatch_after setting ]